rename MergedResultMetaData to EncryptorMetaData
